requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python
numpy
pyyaml
matplotlib
tqdm
